Several underlying non-functional changes towards Metal argument buffer support.

+++ MoltenVK/MoltenVK/GPUObjects/MVKDevice.h
@@ -369,7 +369,7 @@ protected:
 	MVKArrayRef<MVKQueueFamily*> getQueueFamilies();
 	void initPipelineCacheUUID();
 	uint32_t getHighestMTLFeatureSet();
-	uint64_t getMoltenVKGitRevision();
+	uint32_t getMoltenVKGitRevision();
 	void populate(VkPhysicalDeviceIDProperties* pDevIdProps);
 	void logGPUInfo();
